A little something to help me deal with users
When I name folders, I prefer not to have spaces and instead use CamelCase for my folder names. But when users name folders, they don't always conform to my preferences. When assigning permissions to a folder, I like to use a group named after the folder but don't want spaces in the group name. Hence I use this function.